Refactor guardian modules and preserve alias overrides
Split guardian logic into focused modules, rename the reusable child session plumbing to review-session terminology, and move the guardian snapshot alongside the moved tests.
Also preserve explicit false values when migrating smart_approvals to guardian_approval.
