+++ codex-rs/core/src/config/config_tests.rs
@@ -5644,6 +5644,42 @@ smart_approvals = true
 }
 
 #[tokio::test]
+async fn smart_approvals_alias_migration_preserves_disabled_profile_override() -> std::io::Result<()>
+{
+    let codex_home = TempDir::new()?;
+    std::fs::write(
+        codex_home.path().join(CONFIG_TOML_FILE),
+        r#"[features]
+guardian_approval = true
+
+[profiles.guardian.features]
+smart_approvals = false
+"#,
+    )?;
+
+    let config = ConfigBuilder::default()
+        .codex_home(codex_home.path().to_path_buf())
+        .fallback_cwd(Some(codex_home.path().to_path_buf()))
+        .harness_overrides(ConfigOverrides {
+            config_profile: Some("guardian".to_string()),
+            ..Default::default()
+        })
+        .build()
+        .await?;
+
+    assert!(!config.features.enabled(Feature::GuardianApproval));
+    assert_eq!(config.features.legacy_feature_usages().count(), 0);
+    assert_eq!(config.approvals_reviewer, ApprovalsReviewer::User);
+
+    let serialized = tokio::fs::read_to_string(codex_home.path().join(CONFIG_TOML_FILE)).await?;
+    assert!(serialized.contains("[profiles.guardian.features]"));
+    assert!(serialized.contains("guardian_approval = false"));
+    assert!(!serialized.contains("smart_approvals"));
+
+    Ok(())
+}
+
+#[tokio::test]
 async fn smart_approvals_alias_migration_preserves_existing_approvals_reviewer()
 -> std::io::Result<()> {
     let codex_home = TempDir::new()?;
