+++ codex-rs/core/src/lib.rs
@@ -43,7 +43,6 @@ mod file_watcher;
 mod flags;
 pub mod git_info;
 mod guardian;
-mod guardian_subagent;
 pub mod instructions;
 pub mod landlock;
 pub mod mcp;
